Display private host and public IPv4 gateway
The script now also displays the private and public IPV4 addresses of
the host it is running on.

The private IPv4 address is read directly from the local gateway network
interface.

The public IPv4 address is determined by sending an A record query for
the "myip.opendns.com" address to the OpenDNS.com public DNS servers
(resolver1.opendns.com, resolver2.opendns.com) which returns the public
internet gateway IP address of the client doing the request.